derive a peru.yaml from vendor.json
Proof of concept. The potential benefit here is that we would be allowed
to lie about the URL of a package. So for example, if we want to fork
golang-crypto into our own GitHub repo, we wouldn't need to
search/replace all the internal imports in that repo. Instead, we just
have the peru.yaml file use our repo URL, but the original import path,
and go build is none the wiser.

Note that the URLs where go repos actually live are not necessarily the
same as what their import paths say. golang.org in particular seems to
be a redirect or something to go.googlesource.com. That took some
figuring out by hand.