original, unmodified .deb cron file from Chrome
Earlier we pulled out a lot of code from here, but we tried to avoid
bringing in the cronjobs, because that felt complicated. Turns out the
Chrome people know what they're doing (duh), and the cronjobs are
required to handle dist upgrades properly. We need to stick closer to
the Chrome approach.

This is a verbatim copy of the Chrome cron file for Debian, so that it's
easy to see what we're changing in the following commit.